# backend/storage/api-docs-admin/api-docs-admin.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Next Starter API — Admin",
        "description": "Documentação da área **Admin** (equipe interna / back-office) do projeto Laravel Next Starter.\n\nEsta documentação é restrita: o acesso à interface Swagger UI e ao JSON da spec exige autenticação HTTP Basic própria da documentação (usuário/senha definidos em `.env`), **independente** do login da API.\n\nApós acessar a documentação, os endpoints em si continuam exigindo o Bearer JWT obtido em `POST /api/admin/login`.",
        "contact": {
            "name": "Equipe de Desenvolvimento"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor da API (relativo ao host onde o Swagger UI está sendo servido)"
        }
    ],
    "paths": {
...
        "/admin/perfil": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Atualizar dados do perfil",
                "description": "Atualiza nome e/ou e-mail do usuário autenticado. Ambos os campos são opcionais (envie apenas o que deseja alterar).",
                "operationId": "c42e9c4bbba816a6bba6e6615b3c3a98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "description": "Deve ser único entre os usuários.",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil atualizado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Usuario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/perfil/senha": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Atualizar senha do perfil",
                "description": "Atualiza a senha do usuário autenticado. Exige a senha atual (validada no servidor) e a nova senha, com política mínima de complexidade.",
                "operationId": "21a33e9c2e1bc05686aaa201d47c05b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "senha_atual",
                                    "senha_nova",
                                    "senha_nova_confirma"
                                ],
                                "properties": {
                                    "senha_atual": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "senha_nova": {
                                        "description": "Mínimo 8 caracteres, com maiúscula, minúscula, número e símbolo. Deve ser diferente da senha atual.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "NovaSenhaForte@123"
                                    },
                                    "senha_nova_confirma": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NovaSenhaForte@123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha atualizada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Usuario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/perfil/avatar": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Atualizar avatar do perfil",
                "description": "Atualiza o avatar do usuário autenticado a partir de uma imagem codificada em Base64 (apenas PNG ou JPG).",
                "operationId": "90834eeb1d97b8d9cc9590ea84792a1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "description": "Conteúdo da imagem (PNG ou JPG) codificado em Base64.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar atualizado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Usuario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/perfil/sessoes": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Listar sessões ativas do perfil",
                "description": "Lista as sessões ativas (dispositivos/navegadores logados) do usuário autenticado.",
                "operationId": "c9c9a286996414ac1d14ae9280c5c70c",
                "responses": {
                    "200": {
                        "description": "Lista de sessões ativas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UsuarioSessao"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/perfil/sessoes/{id}/encerrar": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Encerrar sessão do perfil",
                "description": "Encerra uma sessão ativa específica do usuário autenticado (ex.: deslogar remotamente de outro dispositivo).",
                "operationId": "9469989d291b9c6684e27e2b75f27838",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/NoContent"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationErrorResponse": {
                "properties": {
                    "errors": {
                        "description": "Formato de erro de validação (422), produzido pelo handler de exceções\nglobal (bootstrap/app.php) a partir de uma ValidationException.",
                        "type": "object",
                        "example": {
                            "campo": [
                                "A mensagem de validação correspondente."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "BusinessErrorResponse": {
                "properties": {
                    "errors": {
                        "properties": {
                            "business": {
                                "description": "Formato padrão de erro de negócio/HTTP (400/401/403/404/500), produzido\npelo mesmo handler global para BusinessException, AuthorizationException,\nModelNotFoundException, HttpExceptionInterface, etc.",
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Descrição do erro de negócio."
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Usuario": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "grupo_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "nome": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "convidado",
                            "ativo",
                            "expirado",
                            "inativo",
                            "bloqueado"
                        ]
                    },
                    "avatar": {
                        "type": "string",
                        "nullable": true
                    },
                    "google2fa_enable": {
                        "type": "boolean"
                    },
                    "google2fa_confirmado_em": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "ultimo_login_em": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "ultimo_ip": {
                        "type": "string",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UsuarioSessao": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "usuario_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "ip": {
                        "type": "string",
                        "nullable": true
                    },
                    "user_agent": {
                        "type": "string",
                        "nullable": true
                    },
                    "browser": {
                        "type": "string",
                        "nullable": true
                    },
                    "plataforma": {
                        "type": "string",
                        "nullable": true
                    },
                    "dispositivo": {
                        "type": "string",
                        "nullable": true
                    },
                    "ativo": {
                        "type": "boolean"
                    },
                    "ultimo_acesso_em": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "logout_em": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "atual": {
                        "description": "Indica se é a sessão da requisição atual.",
                        "type": "boolean"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "ValidationError": {
                "description": "Erro de validação dos campos enviados.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationErrorResponse"
                        }
                    }
                }
            },
...
            "Unauthorized": {
                "description": "Não autenticado: token ausente, inválido, expirado, ou sessão associada ao token não está mais ativa.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BusinessErrorResponse"
                        }
                    }
                }
            },
...
            "NotFound": {
                "description": "Registro não encontrado.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BusinessErrorResponse"
                        }
                    }
                }
            },
...
            "NoContent": {
                "description": "Operação realizada com sucesso, sem conteúdo de retorno."
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via **JSON Web Token (JWT)**.\n\n1. Faça login em `POST /api/login` (Private) ou `POST /api/admin/login` (Admin).\n2. Envie o token retornado no header `Authorization: Bearer {token}` nas demais requisições.\nO token carrega um `session_id`, validado a cada requisição contra a sessão ativa do usuário (tabela de sessões).",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}